+++ lib/instream/cluster/database.ex
@@ -1,39 +1,34 @@
 defmodule Instream.Cluster.Database do
-  @moduledoc """
-  Database administration helper.
-  """
+  @moduledoc false
 
-  alias Instream.Query.Builder
-  alias Instream.Validate
+  alias Instream.Admin
 
-  @doc """
-  Returns a query to create a database.
-
-  Options:
-
-  - `:if_not_exists` - setting to `true` appends "IF NOT EXISTS" to the query.
-  """
+  @doc false
   @spec create(String.t, Keyword.t) :: Builder.t
   def create(database, opts \\ []) do
-    database
-    |> Validate.database!
-    |> Builder.create_database()
-    |> Builder.if_not_exists(opts[:if_not_exists] || false)
+    warn()
+    Admin.Database.create(database, opts)
   end
 
-  @doc """
-  Returns a query to drop a database.
-  """
+  @doc false
   @spec drop(String.t) :: Builder.t
   def drop(database) do
-    database
-    |> Validate.database!
-    |> Builder.drop_database()
+    warn()
+    Admin.Database.drop(database)
   end
 
-  @doc """
-  Returns a query to list databases.
-  """
+  @doc false
   @spec show() :: Builder.t
-  def show(), do: Builder.show(:databases)
+  def show() do
+    warn()
+    Admin.Database.show()
+  end
+
+
+  defp warn do
+    IO.write :stderr, "warning: Instream.Cluster.Database has been renamed" <>
+                      " to Instream.Admin.Database. This module will be" <>
+                      " removed in an upcoming release\n" <>
+                      Exception.format_stacktrace
+  end
 end