+++ lib/instream/cluster/diagnostics.ex
@@ -1,13 +1,20 @@
 defmodule Instream.Cluster.Diagnostics do
-  @moduledoc """
-  Diagnostics query helper.
-  """
+  @moduledoc false
 
-  alias Instream.Query.Builder
+  alias Instream.Admin
 
-  @doc """
-  Returns a query to retrieve server diagnostics.
-  """
+  @doc false
   @spec show() :: Builder.t
-  def show(), do: Builder.show(:diagnostics)
+  def show() do
+    warn()
+    Admin.Diagnostics.show()
+  end
+
+
+  defp warn do
+    IO.write :stderr, "warning: Instream.Cluster.Diagnostics has been" <>
+                      " renamed to Instream.Admin.Diagnostics. This module" <>
+                      " will be removed in an upcoming release\n" <>
+                      Exception.format_stacktrace
+  end
 end