+++ test/instream/deprecations/cluster_diagnostics_test.exs
@@ -1,0 +1,15 @@
+defmodule Instream.Deprecations.ClusterDiagnosticsTest do
+  use ExUnit.Case, async: false
+
+  import ExUnit.CaptureIO
+
+  alias Instream.Cluster.Diagnostics
+
+  test "show" do
+    stderr = capture_io :stderr, fn ->
+      Diagnostics.show()
+    end
+
+    assert String.contains?(stderr, "renamed")
+  end
+end