+++ test/instream/response_test.exs
@@ -1,7 +1,7 @@
 defmodule Instream.ResponseTest do
   use ExUnit.Case, async: true
 
-  alias Instream.Cluster.Database
+  alias Instream.Admin.Database
   alias Instream.Response
   alias Instream.TestHelpers.Connection